Former-commit-id: 1115b7388f8b332048879449e26c1d3ce77e86e8

+++ processm.services/src/main/kotlin/processm/services/api/DataStoresApi.kt
@@ -73,7 +73,7 @@ fun Route.DataStoresApi() {
                 DataStore(
                     dataStore.name,
                     dataStore.id.value,
-                    dataStoreSize.toInt(),
+                    dataStoreSize,
                     dataStore.creationDate
                 )
             )
